Use a marker to detect SQOA file
In this version, all new operations use the "RUN" opcode.

The marker is 48 '0' which cannot occur in a valid QOI file
because the index is initialised with zeros so only index 0
could match the pixel (r=0 g=0 b=0 a=0) at the start.

This version loses to QOI slightly in the screenshot_game
dataset. It still wins in the icons datasets.